+++ CHANGELOG.md
@@ -12,7 +12,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Support for arbitrary (not based on a specific DicomTag) columns in image table templates
 
 ### Changed
+
 - Updated to latest version of FAnsiSql (0.9.2)
+
+### Fixed
+
+- DicomTypeTranslaterWriter: Fixed parsing of private sequence elements when the private creator is unknown
+
 
 ## [2.0.0]